Use set-cookie value as cookieJar directly
As opposed to appending individual cookies each time which causes duplicate cookies to be added to requests until they are too long to send.

Note that there was some logic initially for excluding certain cookies. This doesn't seem to be required for EBI to work, requests work correctly either way. The removed snippet was added when attempting to support self-signed SSL and I have tested that self-signed SSL still works after this change.